feat: AI insights panel displays post-call insights for the agent's last call
- Add lastCallUniqueId to activeCall store, seeded from src_callid on
  incoming call and overwritten with the accurate CHAN_UNIQUEID when AMI
  enrichment runs. Survives wrapup so it is available on case-creation page.

- Extract fetch logic into useAiInsightsFetcher composable with module-level
  singletons for dedup (_fetchedCallIds) and concurrency (_aiFetchInProgress),
  shared across DefaultLayout, CaseCreate and CaseSingleFormView.

- DefaultLayout now includes src_callid and lastCallUniqueId in the callIds
  match set, making AI notification matching resilient when the AMI WebSocket
  is unavailable or the agent extension is not configured.

- CaseCreate and CaseSingleFormView both run a catch-up fetch on mount so
  insights that arrived before the agent navigated to case-creation are
  still loaded and shown in the panel.

Closes #528
